Adds car turn rate control mode
Implements a new car mode that allows direct control of the car's turning rate.

This allows the user to specify a target turning rate in degrees per second, which the car will then attempt to maintain using a PID controller on the gyro Z axis.

Also updates telemetry to output linear acceleration.